// Angular_project/src/app/models/UtilisateurTest.ts
export class UtilisateurTest {
    id: number;
    utilisateur: number;
    test: number;
    score: number;
    sate: Date;
    flagechecsucces: number;
    nb_reponses_correctes: number;
    nb_questions_non_repondues: number;
    nb_questions_repondues: number;
    nb_reponse_incorrectes: number;
    num_essai: number;

   
}